# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/imdb": {
            "get": {
                "description": "Returns a paginated list of all movies from the IMDB database.\nUse the \"page\" and \"page_size\" query parameters to control pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "List all movies (paginated)",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page (default: 10, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of movies",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/imdb/genre": {
            "get": {
                "description": "Returns a paginated list of movies filtered by genre.\nUse the \"g\" query parameter to specify the genre (e.g., \"Action\", \"Drama\", \"Comedy\").",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Filter movies by genre",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"Action\"",
                        "description": "Genre to filter by",
                        "name": "g",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page (default: 10, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtered movies by genre",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request (missing genre)",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Not implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/imdb/search": {
            "get": {
                "description": "Searches movies where the title or overview contains the given keyword (case-insensitive).\nSupports pagination with \"page\" and \"page_size\" parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Search movies by keyword",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"Shawshank\"",
                        "description": "Search keyword",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page (default: 10, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated search results",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request (missing query)",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/imdb/top": {
            "get": {
                "description": "Returns the top N rated movies from the IMDB database.\nUse the \"n\" query parameter to control how many movies to return (default: 10).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get top rated movies",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of top movies to return (default: 10)",
                        "name": "n",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Top rated movies",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/imdb/year": {
            "get": {
                "description": "Returns a paginated list of movies filtered by release year.\nUse the \"y\" query parameter to specify the year (e.g., \"1994\", \"2020\").",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Filter movies by release year",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1994,
                        "description": "Release year to filter by",
                        "name": "y",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page (default: 10, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtered movies by year",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request (missing or invalid year)",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Not implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/imdb/{title}": {
            "get": {
                "description": "Returns a single movie matching the exact title specified in the URL path.\nThe title is URL-decoded automatically by Gin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get a movie by its exact title",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"The Shawshank Redemption\"",
                        "description": "Movie title (URL encoded)",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie found",
                        "schema": {
                            "$ref": "#/definitions/dto.IMDBMovieResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request (missing title)",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Movie not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "movie not found"
                }
            }
        },
        "dto.IMDBMovieResponse": {
            "type": "object",
            "properties": {
                "certificate": {
                    "type": "string",
                    "example": "R"
                },
                "director": {
                    "type": "string",
                    "example": "Frank Darabont"
                },
                "genre": {
                    "type": "string",
                    "example": "Drama"
                },
                "gross": {
                    "type": "string",
                    "example": "$28,341,469"
                },
                "imdb_rating": {
                    "type": "string",
                    "example": "9.3"
                },
                "meta_score": {
                    "type": "string",
                    "example": "80"
                },
                "no_of_votes": {
                    "type": "string",
                    "example": "2343110"
                },
                "overview": {
                    "type": "string",
                    "example": "Two imprisoned men bond over a number of years, finding solace and eventual redemption through acts of common decency."
                },
                "poster_link": {
                    "type": "string",
                    "example": "https://m.media-amazon.com/images/M/MV5BMDFkYTc0MGEtZmNhMC00ZDIzLWFmNTEtODM1ZmRlYWMwMWFmXkEyXkFqcGdeQXVyMTMxODk2OTU@._V1_UX67_CR0,0,67,98_AL_.jpg"
                },
                "released_year": {
                    "type": "string",
                    "example": "1994"
                },
                "runtime": {
                    "type": "string",
                    "example": "142 min"
                },
                "series_title": {
                    "type": "string",
                    "example": "The Shawshank Redemption"
                },
                "stars": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Tim Robbins",
                        "Morgan Freeman",
                        "Bob Gunton"
                    ]
                }
            }
        },
        "dto.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.IMDBMovieResponse"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 10
                },
                "total_count": {
                    "type": "integer",
                    "example": 1000
                },
                "total_pages": {
                    "type": "integer",
                    "example": 100
                }
            }
        }
    }
}